u8

+++ rust/cloud-storage/macro_event_broker/examples/example_event.rs
@@ -56,7 +56,7 @@ pub enum ExampleTopicEvent {
 impl TopicEvent for ExampleTopicEvent {
     type Topic = MacroExampleTopic;
 
-    fn schema_version(&self) -> u16 {
+    fn schema_version(&self) -> u8 {
         1
     }
 }
